Trying to implement hexadeciaml theme color (show only)
For now, some colors will not change because the color in GoodTwitter2 (opt_display_userColor) has been changed.
Also, changing user color is not implemented.

We can still use profile_link_color (a kind of user object) to get or change the theme color in hexadecimal.
However, seems profile_link_color cannot be retrieved by requestUser(), so I decided to use users/show to retrieve it (requestUserAlt()).

P.S. In Twitter for Android, if a user doesn't have a header image, profile_link_color will be displayed in place of it.